refactor: move book_balance_live
