--- apps/app_web/lib/app_web/live/balance_live/show.html.heex
@@ -1,31 +1,0 @@
-<section>
-  <.heading level={:section} class="mt-4">
-    <%= gettext("How to balance?") %>
-  </.heading>
-  <.list>
-    <%= if @transactions_error? do %>
-      <.list_item>
-        <%= gettext("An error occured while balancing the transactions") %>
-      </.list_item>
-    <% else %>
-      <%= if Enum.empty?(@transactions) do %>
-        <.list_item>
-          <%= gettext("The transactions are balanced already!") %>
-        </.list_item>
-      <% else %>
-        <.list_item :for={transaction <- @transactions}>
-          <div class="grow">
-            <span class="font-bold">
-              <%= transaction.from.display_name %>
-            </span>
-            <span><%= gettext("gives") %></span>
-            <span class="font-bold">
-              <%= transaction.to.display_name %>
-            </span>
-          </div>
-          <%= transaction.amount %>
-        </.list_item>
-      <% end %>
-    <% end %>
-  </.list>
-</section>
